Move Object types into their own module. Cleaning up the git module.
